requirements.txt
streamlit==1.35.0
face_recognition==1.3.0
opencv-python==4.9.0.80
web3==6.15.1
numpy==1.26.4
Pillow==10.3.0
